+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   - Bumped discord.py to version 2.4.0
   - Migrated from depreciated logs_from to history
   - Added some network error handling
+  - Added basic rate handling
+  - Added basic reaction support
 
 ## [4.0.1] Unreleased